+++ src/features/article/model/useArticle.ts
@@ -1,0 +1,25 @@
+import { useState, useEffect } from 'react';
+import { Article } from './types';
+import { articleApi } from '../api/articleApi';
+
+export function useArticle(id: string) {
+  const [article, setArticle] = useState<Article | null>(null);
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchArticle = async () => {
+      try {
+        const data = await articleApi.getArticle(id);
+        setArticle(data);
+      } catch (error) {
+        console.error('Failed to fetch article:', error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
+    fetchArticle();
+  }, [id]);
+
+  return { article, isLoading };
+}
